+++ CHANGELOG.md
@@ -7,6 +7,10 @@ Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image 
 ## [1.5.1] - 2021-03-07
 
 - Use shared Mega-Linter configuration
+- New package commands
+  - **hardis:package:install**
+  - **hardis:package:version:create**
+  - **hardis:package:version:list**
 
 ## [1.5.0] 2021-03-05